+++ coz-shopping/src/components/Home/Main.tsx
@@ -3,7 +3,7 @@ import axios from "axios";
 import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { RootState } from "../../store/rootStore";
-import { BookMarkStar } from "../ProductList/ProductList";
+import { BookMarkStar } from "../ProductList/ProductListStyle";
 import {
   addBookMarkedProducts,
   deleteBookMarkedProduct,
