+++ coz-shopping/src/components/ProductList/ProductList.tsx
@@ -20,48 +20,13 @@ import {
   deleteBookMarkedProduct,
 } from "../../store/bookMarkStore";
 import { IItem } from "../Home/Main";
-
-export const ProductListMainWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: flex-start;
-  height: 84vh;
-`;
-
-export const Section = styled.section`
-  margin: 0 20px 0px;
-  padding-bottom: 100px;
-`;
-
-interface IImageProps {
-  id: any;
-}
-
-const enum BookMarkIcon {
-  onIcon = "/image/bookmark_on.jpg",
-  offIcon = "/image/bookmark_off.jpg",
-}
-
-export const BookMarkStar = styled.div<IImageProps>`
-  background-image: ${(props: IImageProps): string => {
-    const bookMarkedProducts = useSelector(
-      (store: RootState) => store.bookMarkedProducts
-    );
-    const isBookMarked = bookMarkedProducts.find(
-      (product: IItem) => product.id === props.id
-    );
-    const imageUrl = isBookMarked ? BookMarkIcon.onIcon : BookMarkIcon.offIcon;
-    return `url(${imageUrl})`;
-  }};
-  border: none;
-  height: 24px;
-  width: 25px;
-  margin-top: -90px;
-  margin-left: 260px;
-  z-index: 0.5;
-  cursor: pointer;
-`;
+import {
+  ProductListMainWrapper,
+  Section,
+  IImageProps,
+  BookMarkIcon,
+  BookMarkStar,
+} from "./ProductListStyle";
 
 function ProductList() {
   const [items, setItems] = useState<IItem[]>([]);
